# swagger-spring/AutoSCFE-1.6.0-swagger.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0.0'
  title: Api Documentation
  termsOfService: urn:tos
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  contact:
      name: Team
      email: team@gruposantander.com
  x-santander-catalogation: 
      bian-landscape-version: "8"
      bian-business-area: “Business Support”
      bian-business-domain: “Document Management and Archive”
      bian-service-domain: "Document Services"  
host: localhost:8080
basePath: /AutoSCFE
securityDefinitions:
  JWTProfile:
    type: "oauth2"
    description: "OAuth Grant type, JWT Profile"
    flow: "implicit"
    scopes:
     sce: "sce"
    authorizationUrl: "https://tbd.santander.es"
    x-tokenIntrospect:
      url: "$(sca-security-url)"
  X-Santander-Client-Id:
    type: "apiKey"
    in: "header"
    name: "X-Santander-Client-Id"
    description: "Client id header."
...
paths:
  /additionalLoanApplication/loan/routing/route:
    post:
      tags:
        - service-route
      summary: Get route
      description: Code information
      operationId: requestRouting
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: Accept
          in: header
          description: The server response
          required: true
          type: string
        - name: Authorization
          in: header
          description: Authorization token. Bearer OAuth2 token
          required: true
          type: string
        - name: Content-Type
          in: header
          description: metadata content type
          required: true
          type: string
        - name: Context
          in: header
          description: include context object
          required: true
          type: string
        - name: X-B3-ParentSpanId
          in: header
          description: 'Position of the parent operation in the trace tree. The value is 64 bits long. value is omitted when the span is the root of the trace tree. '
          required: false
          type: string
        - name: X-B3-Sampled
          in: header
          description: 'Sampling decision. Sampling is a mechanism to reduce the volume of data in the tracing system. In B3, sampling applies consistently per-trace: once the sampling decision is made, the same value must be consistently sent downstream. This means that either all or no spans share a trace ID. The possible values are 0 = Deny 1 = Accept d = Debug'
          required: false
          type: string
        - name: X-B3-SpanId
          in: header
          description: Position of the current operation in the trace tree. The value is 64 bits long. Do not integererpret the value it may or may not be derived from the value of the TraceId.
          required: false
          type: string
        - name: X-B3-TraceId
          in: header
          description: Overall ID of the trace, shared by every span in the trace. The value is 64 or 128 bits long.
          required: false
          type: string
        - name: X-Santander-Client-Id
          in: header
          description: Client Id header
          required: true
          type: string
        - in: body
          name: infoRoutingData
          description: infoRoutingData
          required: true
          schema:
            $ref: '#/definitions/InfoRoutingData'
      responses:
        '200':
          description: 200 OK
          schema:
            $ref: '#/definitions/RoutingResponse'
        '201':
          description: Created
        '400':
          description: Bad Request
          schema:
            type: object
            additionalProperties:
              type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Error Server
      security:
        - JWTProfile:
            - sce
      deprecated: false
...
definitions:
  Address:
    type: object
    properties:
      city:
        type: string
      line1:
        type: string
      line2:
        type: string
      postalCode:
        type: string
      stateOrProvince:
        type: string
    title: Address
  ApplicationArea:
    type: object
    properties:
      sender:
        $ref: '#/definitions/Sender'
    title: ApplicationArea
  ApplicationFunnel:
    type: object
    properties:
      URL:
        type: string
      lifeSpan:
        type: integer
        format: int64
    title: ApplicationFunnel
...
  FinancialDataRequested:
    type: object
    properties:
      applicationType:
        type: string
      damageInsurance:
        type: string
      downPayment:
        type: integer
        format: int32
      employmentInsurance:
        type: string
      lifeInsurance:
        type: string
      paymentAmount:
        type: number
        format: double
      productID:
        type: string
      termInMonths:
        type: integer
        format: int32
    title: FinancialDataRequested
  InfoRoutingData:
    type: object
    properties:
      applicationArea:
        $ref: '#/definitions/ApplicationArea'
      financialDataRequested:
        $ref: '#/definitions/FinancialDataRequested'
      primaryApplicant:
        $ref: '#/definitions/PrimaryApplicant'
      requestMetadata:
        $ref: '#/definitions/RequestMetadata'
      userRedirectURL:
        type: string
      vehicleInfo:
        $ref: '#/definitions/VehicleInfo'
    title: InfoRoutingData
...
  Odometer:
    type: object
    properties:
      UOM:
        type: string
      numUnits:
        type: integer
        format: int32
    title: Odometer
  PrimaryApplicant:
    type: object
    properties:
      address:
        $ref: '#/definitions/Address'
      applicantType:
        type: string
      customerType:
        type: string
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      phone:
        type: string
    title: PrimaryApplicant
...
  Registration:
    type: object
    properties:
      firstDate:
        type: string
      plate:
        type: string
    title: Registration
  RequestMetadata:
    type: object
    properties:
      conversationID:
        type: string
      countryCode:
        type: string
      currency:
        type: string
      referenceNumber:
        type: string
    title: RequestMetadata
  RoutingResponse:
    type: object
    properties:
      applicationFunnel:
        $ref: '#/definitions/ApplicationFunnel'
      financeID:
        type: string
    title: RoutingResponse
  Sender:
    type: object
    properties:
      ID:
        type: string
      language:
        type: string
      name:
        type: string
    title: Sender
...
  VehicleInfo:
    type: object
    properties:
      Odometer:
        $ref: '#/definitions/Odometer'
      VIN:
        type: string
      category:
        type: string
      dealerOwner:
        type: string
      manufacturer:
        type: string
      model:
        type: string
      modelID:
        type: string
      modelYear:
        type: string
      pictureURL:
        type: string
      registration:
        $ref: '#/definitions/Registration'
      saleClass:
        type: string
      trim:
        type: string
      vehiclePriceBreakdown:
        $ref: '#/definitions/VehiclePriceBreakdown'
    title: VehicleInfo
  VehiclePriceBreakdown:
    type: object
    properties:
      totalVehiclePrice:
        type: integer
        format: int32
      vehiclePrice:
        type: integer
        format: int32
    title: VehiclePriceBreakdown
...